置換 "Float" -> "Double" (一部除外) #2

+++ examples/char-rnn/main.rs
@@ -24,13 +24,13 @@ fn sample(data: &TextData, lstm: &LSTM, linear: &Linear, device: Device) -> Stri
     let mut last_label = 0i64;
     let mut result = String::new();
     for _index in 0..SAMPLING_LEN {
-        let input = Tensor::zeros(&[1, labels], (Kind::Float, device));
+        let input = Tensor::zeros(&[1, labels], (Kind::Double, device));
         let _ = input.narrow(1, last_label, 1).fill_(1.0);
         state = lstm.step(&input, &state);
         let sampled_y = linear
             .forward(&state.h())
             .squeeze_dim(0)
-            .softmax(-1, Kind::Float)
+            .softmax(-1, Kind::Double)
             .multinomial(1, false);
         last_label = i64::from(sampled_y);
         result.push(data.label_to_char(last_label))
