+++ examples/gan/main.rs
@@ -72,7 +72,7 @@ fn discriminator(p: nn::Path) -> impl nn::ModuleT {
 
 fn mse_loss(x: &Tensor, y: &Tensor) -> Tensor {
     let diff = x - y;
-    (&diff * &diff).mean(Kind::Float)
+    (&diff * &diff).mean(Kind::Double)
 }
 
 // Generate a 2D matrix of images from a tensor with multiple images.
@@ -106,7 +106,7 @@ pub fn main() -> Result<()> {
         images
             .index_select(0, &index)
             .to_device(device)
-            .to_kind(Kind::Float)
+            .to_kind(Kind::Double)
             / 127.5
             - 1.
     };
@@ -136,8 +136,8 @@ pub fn main() -> Result<()> {
                 .copy()
                 .detach()
                 .apply_t(&discriminator, true);
-            mse_loss(&y_pred, &(y_pred_fake.mean(Kind::Float) + 1))
-                + mse_loss(&y_pred_fake, &(y_pred.mean(Kind::Float) - 1))
+            mse_loss(&y_pred, &(y_pred_fake.mean(Kind::Double) + 1))
+                + mse_loss(&y_pred_fake, &(y_pred.mean(Kind::Double) - 1))
         };
         opt_d.backward_step(&discriminator_loss);
 
@@ -150,8 +150,8 @@ pub fn main() -> Result<()> {
             let y_pred_fake = rand_latent()
                 .apply_t(&generator, true)
                 .apply_t(&discriminator, true);
-            mse_loss(&y_pred, &(y_pred_fake.mean(Kind::Float) - 1))
-                + mse_loss(&y_pred_fake, &(y_pred.mean(Kind::Float) + 1))
+            mse_loss(&y_pred, &(y_pred_fake.mean(Kind::Double) - 1))
+                + mse_loss(&y_pred_fake, &(y_pred.mean(Kind::Double) + 1))
         };
         opt_g.backward_step(&generator_loss);
