+++ examples/jit/main.rs
@@ -21,7 +21,7 @@ pub fn main() -> Result<()> {
     let output = image
         .unsqueeze(0)
         .apply(&model)
-        .softmax(-1, tch::Kind::Float);
+        .softmax(-1, tch::Kind::Double);
 
     // Print the top 5 categories for this image.
     for (probability, class) in imagenet::top(&output, 5).iter() {
