+++ examples/pretrained-models/main.rs
@@ -64,7 +64,7 @@ pub fn main() -> Result<()> {
     // Apply the forward pass of the model to get the logits.
     let output = net
         .forward_t(&image.unsqueeze(0), /* train= */ false)
-        .softmax(-1, tch::Kind::Float); // Convert to probability.
+        .softmax(-1, tch::Kind::Double); // Convert to probability.
 
     // Print the top 5 categories for this image.
     for (probability, class) in imagenet::top(&output, 5).iter() {
