+++ src/nn/init.rs
@@ -23,27 +23,27 @@ pub fn f_init(i: Init, dims: &[i64], device: Device) -> Result<Tensor, TchError>
         Init::Const(cst) => {
             // Optimize the case for which a single C++ code can be done.
             if cst == 0. {
-                Tensor::f_zeros(dims, (Kind::Float, device))
+                Tensor::f_zeros(dims, (Kind::Double, device))
             } else if (cst - 1.).abs() <= std::f64::EPSILON {
-                Tensor::f_ones(dims, (Kind::Float, device))
+                Tensor::f_ones(dims, (Kind::Double, device))
             } else {
-                Tensor::f_ones(dims, (Kind::Float, device)).map(|t| t * cst)
+                Tensor::f_ones(dims, (Kind::Double, device)).map(|t| t * cst)
             }
         }
         Init::Uniform { lo, up } => {
-            Tensor::f_zeros(dims, (Kind::Float, device))?.f_uniform_(lo, up)
+            Tensor::f_zeros(dims, (Kind::Double, device))?.f_uniform_(lo, up)
         }
         Init::Randn { mean, stdev } => {
             if mean == 0. && (stdev - 1.).abs() <= std::f64::EPSILON {
-                Tensor::f_randn(dims, (Kind::Float, device))
+                Tensor::f_randn(dims, (Kind::Double, device))
             } else {
-                Tensor::f_randn(dims, (Kind::Float, device)).map(|t| t * stdev + mean)
+                Tensor::f_randn(dims, (Kind::Double, device)).map(|t| t * stdev + mean)
             }
         }
         Init::KaimingUniform => {
             let fan_in: i64 = dims.iter().skip(1).product();
             let bound = (1.0 / fan_in as f64).sqrt();
-            Tensor::f_zeros(dims, (Kind::Float, device))?.f_uniform_(-bound, bound)
+            Tensor::f_zeros(dims, (Kind::Double, device))?.f_uniform_(-bound, bound)
         }
     }
 }
