+++ tests/serialization_tests.rs
@@ -20,7 +20,7 @@ fn save_and_load_multi() {
     assert_eq!(named_tensors.len(), 2);
     assert_eq!(named_tensors[0].0, "pi");
     assert_eq!(named_tensors[1].0, "e");
-    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Float)), 57);
+    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Double)), 57);
 }
 
 #[test]
@@ -33,7 +33,7 @@ fn save_and_load_npz() {
     assert_eq!(named_tensors.len(), 2);
     assert_eq!(named_tensors[0].0, "pi");
     assert_eq!(named_tensors[1].0, "e");
-    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Float)), 57);
+    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Double)), 57);
 }
 
 #[test]
@@ -47,7 +47,7 @@ fn save_and_load_npz_half() {
     assert_eq!(named_tensors.len(), 2);
     assert_eq!(named_tensors[0].0, "pi");
     assert_eq!(named_tensors[1].0, "e");
-    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Float)), 57);
+    assert_eq!(i64::from(&named_tensors[1].1.sum(tch::Kind::Double)), 57);
 }
 
 #[test]
